+++ test/functional/test.spec.js
@@ -1,0 +1,1 @@
+require('./specs/validator.spec.js')